// packages/workflows/src/sgp22/tasks.d.ts
import { WorkflowPorts } from '../base/types';
export declare const tasks: {
    initialize: (ports: WorkflowPorts) => import("xstate").PromiseActorLogic<void, import("xstate").NonReducibleUnknown, import("xstate").EventObject>;
    authenticate: (ports: WorkflowPorts) => import("xstate").PromiseActorLogic<{
        transactionId: string;
    }, import("xstate").NonReducibleUnknown, import("xstate").EventObject>;
    downloadProfile: (ports: WorkflowPorts) => import("xstate").PromiseActorLogic<Buffer<ArrayBuffer>, {
        transactionId: string;
    }, import("xstate").EventObject>;
    installSegment: (ports: WorkflowPorts) => import("xstate").PromiseActorLogic<void, {
        apdu: string;
    }, import("xstate").EventObject>;
    getProfilesInfo: (ports: WorkflowPorts) => import("xstate").PromiseActorLogic<{
        success: boolean;
        data?: import("@unuko/core").RAPDU;
        status?: import("@unuko/core").ChipStatus;
        error?: import("@unuko/core").TransportError;
    }, import("xstate").NonReducibleUnknown, import("xstate").EventObject>;
    manageProfile: (ports: WorkflowPorts) => import("xstate").PromiseActorLogic<void, {
        iccid: string;
        action: "enable" | "disable" | "delete";
    }, import("xstate").EventObject>;
    listNotifications: (ports: WorkflowPorts) => import("xstate").PromiseActorLogic<{
        success: boolean;
        data?: import("@unuko/core").RAPDU;
        status?: import("@unuko/core").ChipStatus;
        error?: import("@unuko/core").TransportError;
    }, import("xstate").NonReducibleUnknown, import("xstate").EventObject>;
    handleNotification: (ports: WorkflowPorts) => import("xstate").PromiseActorLogic<void, {
        seqNumber: string;
    }, import("xstate").EventObject>;
    parseNotificationsInfo: (buffer: Buffer) => {
        seqNumber: string | undefined;
        event: string;
    }[];
    parseProfilesInfo: (buffer: Buffer) => {
        iccid: string | undefined;
        name: string;
        status: string;
    }[];
    segmentBPP: (bpp: Buffer) => string[];
    logEvent: (ports: WorkflowPorts, description: string, payload?: any) => Promise<void>;
    logEventInvoke: (ports: WorkflowPorts) => import("xstate").PromiseActorLogic<void, {
        description: string;
        payload?: any;
    }, import("xstate").EventObject>;
    registerSubscriber: (ports: WorkflowPorts) => import("xstate").PromiseActorLogic<unknown, {
        iccid: string;
    }, import("xstate").EventObject>;
    enableConnectivity: (ports: WorkflowPorts) => import("xstate").PromiseActorLogic<void, import("xstate").NonReducibleUnknown, import("xstate").EventObject>;
};